+++ packages/iov-bns/types/context.d.ts
@@ -1,4 +1,4 @@
-import { Address, BcpCoin, BcpTicker, ChainId, ConfirmedTransaction, OpenSwap, SwapOfferTransaction } from "@iov/bcp";
+import { Address, Amount, BcpTicker, ChainId, ConfirmedTransaction, OpenSwap, SwapOfferTransaction } from "@iov/bcp";
 import * as codecImpl from "./generated/codecimpl";
 import { Keyed } from "./types";
 /**
@@ -13,15 +13,13 @@ export interface ChainData {
 /** Like BCP's Account but with no pubkey. Keep compatible to Account! */
 export interface WalletData {
     readonly address: Address;
-    readonly balance: ReadonlyArray<BcpCoin>;
+    readonly balance: ReadonlyArray<Amount>;
 }
 export declare class Context {
     private readonly chainData;
     constructor(chainData: ChainData);
     wallet(acct: codecImpl.namecoin.IWallet & Keyed): WalletData;
-    coin(coin: codecImpl.coin.ICoin): BcpCoin;
     /** Decode within a Context to have the chain ID available */
     decodeOpenSwap(swap: codecImpl.escrow.Escrow & Keyed): OpenSwap;
     swapOfferFromTx(confirmed: ConfirmedTransaction<SwapOfferTransaction>): OpenSwap;
-    private amountToCoin;
 }